refactor: change package setup to use vite
