+++ packages/svelte-vega/package.json
@@ -16,47 +16,58 @@
   },
   "type": "module",
   "main": "dist/index.js",
-  "module": "dist/index.js",
-  "svelte": "dist/index.js",
-  "types": "src/index.ts",
+  "module": "./dist/index.js",
+  "svelte": "./dist/index.js",
+  "types": "./dist/index.d.ts",
   "scripts": {
     "prebuild": "rm -rf ./dist",
-    "build": "npm run build:js && npm run build:svelte",
-    "build:js": "tsc -p . --outDir dist/ --rootDir src/",
-    "build:svelte": "svelte-strip strip src/ dist",
-    "validate": "svelte-check --ignore dist",
-    "prepublishOnly": "npm run build",
+    "build": "npm run package",
+    "package": "svelte-kit sync && svelte-package && publint",
+    "prepublishOnly": "yarn package",
+    "check": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json",
+    "lint": "prettier --plugin-search-dir . --check . && eslint .",
+    "format": "prettier --plugin-search-dir . --write .",
     "semantic-release": "semantic-release"
+  },
+  "exports": {
+    ".": {
+      "types": "./dist/index.d.ts",
+      "svelte": "./dist/index.js"
+    }
   },
   "files": [
     "src",
-    "dist"
+    "dist",
+    "!dist/**/*.test.*",
+    "!dist/**/*.spec.*"
   ],
   "dependencies": {
     "fast-deep-equal": "^3.1.3",
     "vega-embed": "^6.21.3"
   },
   "devDependencies": {
-    "@rollup/plugin-commonjs": "^24.0.1",
-    "@rollup/plugin-json": "^6.0.0",
-    "@rollup/plugin-node-resolve": "^15.0.1",
-    "@rollup/plugin-typescript": "^11.0.0",
-    "@tsconfig/svelte": "^4.0.1",
-    "tslib": "^2.5.0",
-    "typescript": "^5.0.3",
-    "rollup": "^3.20.2",
-    "rollup-plugin-css-only": "^4.3.0",
-    "rollup-plugin-execute": "^1.1.1",
-    "rollup-plugin-svelte": "^7.1.4",
-    "svelte": "^3.58.0",
-    "svelte-check": "^3.1.4",
-    "svelte-preprocess": "^5.0.3",
-    "svelte-strip": "^2.0.0"
+		"@sveltejs/adapter-auto": "^2.0.0",
+		"@sveltejs/kit": "^1.5.0",
+		"@sveltejs/package": "^2.0.0",
+		"@typescript-eslint/eslint-plugin": "^5.45.0",
+		"@typescript-eslint/parser": "^5.45.0",
+		"eslint": "^8.28.0",
+		"eslint-config-prettier": "^8.5.0",
+		"eslint-plugin-svelte3": "^4.0.0",
+		"prettier": "^2.8.0",
+		"prettier-plugin-svelte": "^2.8.1",
+		"publint": "^0.1.9",
+		"svelte": "^3.54.0",
+		"svelte-check": "^3.0.1",
+		"tslib": "^2.4.1",
+		"typescript": "^5.0.0",
+		"vite": "^4.2.0"
   },
-  "peerDependencies": {
+	"peerDependencies": {
+		"svelte": "^3.54.0",
     "vega": "*",
     "vega-lite": "*"
-  },
+	},
   "publishConfig": {
     "access": "public"
   }
