+++ packages/svelte-vega/src/lib/constants.ts
