+++ packages/svelte-vega/src/lib/index.ts
