+++ packages/svelte-vega/src/lib/types.ts
