+++ packages/svelte-vega/src/lib/utils.ts
